requirements: fix to point to upstream
As a temporary measure, we were pointing to Vikas' fork for kuryr-lib
since he did most of the split work. Kuryr-lib is starting to take good
shape and it is necessary for community guidelines and for devstack
correct operation that we point to OpenStack's own kuryr-lib
(openstack/kuryr) repository.

Additionally, I manually published a development package to pypi for
kuryr-lib so that when devstack's pip install of kuryr-libnetwork tries
to find the kuryr-lib version in pypi, it finds one.

In the future, once kuryr-lib stabilizes, it would be good to operate
like Neutron does with Neutron lib, i.e., point to a specific kuryr-lib
release in requirements. Until then, `pip install -r requirements` will
install from master and `pip install .` will find the pypi wheel.

Change-Id: If9b5da3a1853a6aa1e5ef177952591684c0e6053
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>