container: Build the container from tree not git
Up until this patch, the Docker container for the Kuryr libnetwork
daemon was part of contrib and was being built by checking out the
latest master from github. That is not very helpful when developers
want to build the container to run their latest changes.

With this change, we move the Docker building to the root of the repo so
that it can use the current state of the repo to build and update the
README.rst to contain the information relative to building and using the
container.

Change-Id: Ie9032d62c73d873a29cabe84c64a19df13fdb749
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>