devstack: Make docker also listen to its socket
In order to have Docker easily usable for the fullstack testing, we made
it listen to the port 2375. This had an unintended and very bothering
side effect in forcing cli usage to have to pass '-H :2375' to every
signle docker command.

It was driving me crazy, so this should make both the normal cli usage
and the fullstack usage work without hassle.

Change-Id: I0d6162937e018664516de406197ca0a464b4ee71
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>