Allow running kuryr-libnetwork with uwsgi
scripts/run_kuryr.sh now understands KURYR_USE_UWSGI. If set to "True" (default)
it would attempt to look for uwsgi executable and run kuryr-libnetwork
with it.
Number of spawned threads and processes can be controlled with
KURYR_UWSGI_THREADS and KURYR_UWSGI_PROCESSES env variables.

If uwsgi is not present or KURYR_USE_UWSGI is set to false the script
operates as usual.

Targets bp: deploy-kuryr-libnetwork-api-in-wsgi
Co-Authored-By: Kirill Zaitsev <k.zaitsev@me.com>

Change-Id: I821a78649ad6f07e4709fa343a0e678e4193a068