unit_tests: do not base our tests on neutronclient
The python-neutronclient module that we were getting our base test class
from was intended to serve for tests that emulate command line input and
output and, as such, it took control of stdin and stdout.

It is inconvenient to have stdin and stdout captured because that
prevents the developers from printing or attaching debuggers when
running unit tests.

Change-Id: I729652d0b5bf8a852950535cce36195397165581
Closes-bug: #1626542
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>