Move interface creation and deletion
The current implementation calls interface create/delete in
NetworkDriver.Join/NetworkDriver.Leave but Docker has been observed to
release the interface to the default namespace in
NetworkDriver.DeleteEndpoint. This would suggest in order to be
compliant with Docker Libnetwork, interface creation and deletion should
be done in NetworkDriver.CreateEndpoint and NetworkDriver.DeleteEndpoint

Please note that while the implementation before this patch works anyway
for vEth pairs, it is not compatible with other interface types (e.g.
IPVlan) as it would be impossible for Kuryr to delete the endpoint until
Docker moves it to the default namespace

This patch moves the current interface creation/deletion from
NetworkDriver.Join/NetworkDriver.Leave to
NetworkDriver.CreateEndpoint/NetworkDriver.DeleteEndpoint

_create_or_update port was refactored to return the neutron_port and
subnets associated with the interface_cidr, the interface_response
message creation has been moved to NetworkDriver.CreateEndpoint

DocImpact
Co-Authored-By: Marco Chiappero <marco.chiappero@intel.com>
Change-Id: I84999989cec34a2a5b5cd2821155f150223c84cc