Driver based model for kuryr-libnetwork
This patch introduces drivers to perform deployment specific functions
for libnetwork. This patch is based on and complementary to the driver
model introduced in kuryr-lib for port binding, by enabling its use in
kuryr-libnetwork.

Initially the following two drivers will be available and included:
* veth: this driver will exclusively make use of the the corresponding
  kuryr-lib veth driver.
* nested: this driver is intended for nested containers and will enable
  the use of ipvlan/macvlan binding drivers.

The driver will be selected by means of a new 'port_driver' option.

NOTE: Due to https://bugs.launchpad.net/kuryr-libnetwork/+bug/1651015
When trying nested ports, it is important to use Kuryr to create the
networks that will be used by Nova to create the instances, otherwise
the nested ports will not be able to be cleaned.

DocImpact
Co-Authored-By: Luis Tomas Bolivar <ltomasbo@redhat.com>
Co-Authored-By: Louise Daly <louise.m.daly@intel.com>
Co-Authored-By: Gary Loughnane <gary.loughnane@intel.com>
Implements: blueprint driver-binding-ipvlan

Change-Id: I380fb017dc97a2cd54a404ec3c8154ecc2df80b8