Fix bug in ipam_release_address
When user deallocates an ip address with the given PoolID,
if we have another cidrs with the same subnetpool prefix, at
the same time the give PoolID subnet had already deleted
before(for example: run fullstack test remove_network,
we firstly call DeleteNetwork, and then ReleaseAddress),
we will hit the exception described in LaunchPad.

This patch always check the subnetpool id if we get subnets.

Change-Id: Ia097c3613841619ed2beae2f9d56ec26bc530f69
Closes-Bug: #1644404