Delete port created by kuryr
In current code, we get the port's subnet by checking the
subnet's subnetpool_id equal with the given PoolID, but in
existing environment, the subnet may lack subnetpool_id,
hence we will not get the port's subnet and leaves the
kuryr-libnetwork port undeleted.

This patch checks the port's device owner and determines
the given POOL's subnet_id and ip_address is equal to
port's fixed_ips attribute.

Change-Id: I736adbb2409b50a5e11d3aa7d31778944cfdeae1
Co-Authored-By: Liping Mao <limao@cisco.com>
Closes-Bug: #1651015