Remove unnecessary checks in CreateNetwork
As of 93db824cf64ae63fde5bb55206b3651763ec9ed3, a Neutron subnetpool
name gets calculated by Kuryr if not provided by libnetwork in the
request.
As a consequence, the following 'if' checks on pool_name are no longer
needed as the value is now either user provided or calculated. This
patch removes them.

Change-Id: I205b0be1105eca4f14b69db33a8ddc4318040ffc
Signed-off-by: Marco Chiappero <marco.chiappero@intel.com>