Add subnetpool_id tag for existing Neutron subnet
In existing OpenStack environment, if we have two Neutron
networks and subnets with same cidrs(this is normal in Neutron),
then create Docker network using one of the existing Neutron network
with same cidr, it will fails when we booting an container.
In current code, we get subnets by cidr, then determine the subnet's
subnetpool_id whether equal with the given PoolID. As we known,
the existing Neutron subnet may missing subnetpool_id, so it will
not get the correct subnet.

This patch adds subnetpool_id tag for existing Neutron subnet.
And in RequestAddress we will distinguish the subnet created by
Kuryr and Neutron. In Neutron case we will determine if the pool_id
in subnet tags or not.

NOTE:
Neutron tags extension for subnet is required by this patch,
otherwise we will give a warning for users.

Change-Id: I9500427c9fb3716976b888096f510ecbbd70ac13
Closes-Bug: #1651318