Ensure subports are deleted after container deletion
There is a disaligned used of the device_owner field in Neutron
and Kuryr which makes both of them changing the device_owner to
different values, with the consequence that something the remaining
value is the one set by Neutron, and thus making the filter to
filter out the port to be removed, leaving it without being
cleaned-up. This is a temporal fix that can be later enhanced
by using tagging instead of device_owner to handle the filtering,
once the tagging support is merged in Neutron.

Change-Id: I42f2e69e6511469ff1ac03c5fec3492200a34e40