Support creating from existing subnetpool
To use this feature, users need to pre-create the subnetpool and
pass the pool name to kuryr. For example:

$ docker network create -d kuryr --ipam-driver=kuryr ... \
    -o neutron.pool.name=testpool \
    --ipam-opt neutron.pool.name=testpool \
    foo

Kuryr will look for the existing subnetpool in neutron instead of
creating a new subnetpool. If the network is deleted, the existing
subnetpool will not be deleted.

NOTE: the option 'neutron.pool.name' was used to specify an custom
name of subnetpool. This commit will the semantic of this option.

Implements: blueprint existing-subnetpool
Change-Id: I73f6c17e0e941ef9dbd20591002fdf96c59e6424