Fix mac address inconsistencies in neutron and docker
To specify the mac address of network interface, --mac-address should be
explicitly specified when running "docker run". However, the mac_address
of the neutron port owned by container is not set when created in
IpamDriver.RequestAddress and updated in NetworkDriver.CreateEndpoint.
This causes the mac addresses in neutron and the docker to be
inconsistent.

This patch updates the mac_address of port in
NetworkDriver.CreateEndpoint if the param is specified. Besides, due to
mac_address of bound ports can not be changed, the set of
'binding:host_id' is moved from IpamDriver.RequestAddress to
NetworkDriver.CreateEndpoint.

Change-Id: I43e891171a7bd7ef30ab02abea4cdca2e5a39105
Closes-Bug: #1664036