Ensure the returned pool has a matched cidr
For ipam_request_pool, if users specify an existing pool, check
if the cidr of requested pool matches the requested cidr. If
unmached, create a new subnetpool and return the ID of the new
subnetpool instead of the ID of the existing pool. The rational
is that ipam_request_address searches neutron subnet with
matched cidr of the pool. An unmatched cidr will lead to failure
on requesting address for the container.

Change-Id: Ic9d9c7dc93c9d8a1861e777e9200fa9a16a404e7
Closes-Bug: #1695678