Make the name of existing net tag unique
In before, the existing net tag is static. This doesn't work well
if there are two docker network creating from the same neutron net.
In this senario, each docker network will tag the neutron net on
creation and untag the neutron net on deletion. Neutron net will
be tag/untag multiple times with the same value which leads to bugs.

This commit enhance the existing net tag. It makes each tag unique
by appending the first 12 chars of the docker network id. If there
are two docker networks, they will have a different tag. Each tag
will be cleanup on the deletion of its corresponsing network.

Change-Id: I4fe8a7e13832202dae026b0b4df788f95c41de3c
Closes-Bug: #1695682