Support specifying existing subnet uuid/name
Currently, we rely on user-provided cidr to select the subnet from which the
port creates. This approach has some limitations, i.e. overlapping cidr. This
BP proposes to introduce driver-specific options (e.g. neutron.subnet.uuid) to
allow users to specify a specific subnet. To identify the right subnetpools
has the same prefixes. The approach is to use tagging, tag subnetpool with the
uuid of the specified subnet.

For e.x - Need both ipam-option *and* option:
    docker network create ... --ipam-opt neutron.subnet.uuid=xxx -o
    neutron.subnet.uuid=xxx (User can pass neutron.subnet.name also)

Change-Id: I01223e9c7668ba083a90d86e007cc5ff16af84e3
Implement: blueprint existing-subnet