Add installation guide for kuryr-libnetwork
This patch proposes an installation guide that covers step-by-step
deployment of kuryr-libnetwork for new users of OpenStack with
sufficient Linux experience.

The instruction is to install kuryr-libnetwork from source since
there is no distro package for kuryr-libnetwork for now. This guide
covers installation in Ubunut as the first stpe. The support of
other OS is future work.

TODO: Configure the CI to publish this guide to:
https://docs.openstack.org/pike/install/

Change-Id: Id4f234068600564fa18e73aba08f6e43893456ee