Don't overwrite device_id of existing port
Right now, we use libnetwork endpoint ID as the device_id of
neutron port. However, if the neutron port is pre-created by users,
users might want to customize the value of device_id. A use case is
to set the device_id as the ID of the container so that users can
list ports belonging to a specific container.

This patch proposes to perserve the device_id of existing neutron
ports if the device_id is not empty.

Change-Id: I9cdfdb0b0db6bf9ab0983144a0d7f0dc5ecbd05f