modify the kuryr workflow document
In the document, it shows that when user input commands like "sudo docker run --net=foo -itd --name=container1 busybox", '/IpamDriver.RequestAddress' will get a POST call which the 'options' is empty, just like this:
{
    "Address": "",
    "PoolID": "941f790073c3a2c70099ea527ee3a6205e037e84749f2c6e8a5287d9c62fd376",
    "Options": {},
}
but I have checked the workflow, and I found it's not true. The Options
will be non-empty, and it will be like "com.docker.network.endpoint.macaddress": "08:22:e0:a8:7d:db".
Fixes: Bug 1774977

Change-Id: Ia4632fb376471893de71b57b49ca0241cf0f60c2