Clone Kuryr repository by TARGET_BRANCH
The variable $TARGET_BRANCH will be set by devstack. It will point
to the right branch that the devstack is based on. For example,
if the devstack is on a stable branch, TARGET_BRANCH will point to
that stable branch. Using this variable will save us the manual
effort to update the branch whenever we cut a release.

Change-Id: I45f93aa328c4e166d131d18c579651e4e96c3254