Handle single ip port on dual-stack
In the case of dual-stack network, users might create a container
with existing neutron port. In before, kuryr assumes the
existing port is a dual-port (with both v4 and v6 addresses)
but this assumption is not always true. In face, it is
possible to create a v4 only port in a dual-stack network and
use it as an existing port.

This commit handle the case that the container is created from
a dual-net with a specified v4-only port. In this case, kuryr
will create a v6 port in ipam_request_address as a place holder.
The v6 port will be removed at network_driver_create_endpoint.

NOTE: the change in kuryr_libnetwork/tests/fullstack/test_ipam.py
is cherry picked from commit 89485dc390f4857464506550c2f4a38ab22ac8ae
for fixing the job 'kuryr-libnetwork-fullstack-py35'

Related-Bug: #1800375
Change-Id: Id988abf1b6560332b18a60d99658a8768d46c343
(cherry picked from commit 68cb29267a1b0db19c31709959c2b2e88501780d)